.history/requirements_20250116223839.txt
psycopg2-binary
pillow
sanic
